Split interpolators into own files
