+++ src/interpolate.rs
@@ -69,12 +69,18 @@
 //! [`resource_tween_system`]: crate::tween::resource_tween_system
 //! [`asset_tween_system`]: crate::tween::asset_tween_system
 
-use std::sync::Arc;
+mod blanket_impl;
+#[cfg(feature = "bevy_sprite")]
+mod sprite;
+mod transform;
+
+pub use transform::*;
+
+#[cfg(feature = "bevy_sprite")]
+pub use sprite::*;
 
 use bevy::prelude::*;
 
-#[cfg(feature = "bevy_sprite")]
-use crate::utils::color_lerp;
 use crate::{tween, BevyTweenRegisterSystems};
 
 /// Alias for an `Interpolator` as a boxed trait object.
@@ -185,47 +191,6 @@ pub trait Interpolator: Send + Sync + 'static {
 //     }
 // }
 
-impl<I> Interpolator for Box<I>
-where
-    I: Interpolator + ?Sized,
-{
-    type Item = I::Item;
-
-    fn interpolate(&self, item: &mut Self::Item, value: f32) {
-        (**self).interpolate(item, value)
-    }
-}
-
-impl<I> Interpolator for &'static I
-where
-    I: Interpolator + ?Sized,
-{
-    type Item = I::Item;
-
-    fn interpolate(&self, item: &mut Self::Item, value: f32) {
-        (**self).interpolate(item, value)
-    }
-}
-
-impl<I> Interpolator for Arc<I>
-where
-    I: Interpolator + ?Sized,
-{
-    type Item = I::Item;
-
-    fn interpolate(&self, item: &mut Self::Item, value: f32) {
-        (**self).interpolate(item, value)
-    }
-}
-
-impl<I: 'static> Interpolator for dyn Fn(&mut I, f32) + Send + Sync + 'static {
-    type Item = I;
-
-    fn interpolate(&self, item: &mut Self::Item, value: f32) {
-        self(item, value)
-    }
-}
-
 /// Default interpolators
 ///
 /// Register type and systems for the following interpolators:
@@ -259,8 +224,8 @@ impl Plugin for DefaultInterpolatorsPlugin {
             .register_type::<tween::ComponentTween<SpriteColor>>();
 
         #[cfg(all(feature = "bevy_sprite", feature = "bevy_asset",))]
-        app.add_tween_systems(tween::asset_tween_system::<ColorMaterial>())
-            .register_type::<tween::AssetTween<ColorMaterial>>();
+        app.add_tween_systems(tween::asset_tween_system::<sprite::ColorMaterial>())
+            .register_type::<tween::AssetTween<sprite::ColorMaterial>>();
     }
 }
 
@@ -293,258 +258,5 @@ impl Plugin for DefaultDynInterpolatorsPlugin {
         app.add_tween_systems(tween::asset_tween_system::<
             BoxedInterpolator<bevy::sprite::ColorMaterial>,
         >());
-    }
-}
-
-// type ReflectInterpolatorTransform = ReflectInterpolator<Transform>;
-
-/// [`Interpolator`] for [`Transform`]'s translation.
-#[derive(Debug, Default, Clone, PartialEq, Reflect)]
-// #[reflect(InterpolatorTransform)]
-pub struct Translation {
-    #[allow(missing_docs)]
-    pub start: Vec3,
-    #[allow(missing_docs)]
-    pub end: Vec3,
-}
-impl Interpolator for Translation {
-    type Item = Transform;
-
-    fn interpolate(&self, item: &mut Self::Item, value: f32) {
-        item.translation = self.start.lerp(self.end, value);
-    }
-}
-
-/// Constructor for [`Translation`]
-pub fn translation(start: Vec3, end: Vec3) -> Translation {
-    Translation { start, end }
-}
-
-/// Constructor for [`Translation`] that's relative to previous value using currying.
-pub fn translation_to(to: Vec3) -> impl Fn(&mut Vec3) -> Translation {
-    move |state| {
-        let start = *state;
-        let end = to;
-        *state = to;
-        translation(start, end)
-    }
-}
-
-/// Constructor for [`Translation`] that's relative to previous value using currying.
-pub fn translation_by(by: Vec3) -> impl Fn(&mut Vec3) -> Translation {
-    move |state| {
-        let start = *state;
-        let end = *state + by;
-        *state += by;
-        translation(start, end)
-    }
-}
-
-/// [`Interpolator`] for [`Transform`]'s rotation using the [`Quat::slerp`] function.
-#[derive(Debug, Default, Clone, PartialEq, Reflect)]
-// #[reflect(InterpolatorTransform)]
-pub struct Rotation {
-    #[allow(missing_docs)]
-    pub start: Quat,
-    #[allow(missing_docs)]
-    pub end: Quat,
-}
-impl Interpolator for Rotation {
-    type Item = Transform;
-
-    fn interpolate(&self, item: &mut Self::Item, value: f32) {
-        item.rotation = self.start.slerp(self.end, value);
-    }
-}
-
-/// Constructor for [`Rotation`]
-pub fn rotation(start: Quat, end: Quat) -> Rotation {
-    Rotation { start, end }
-}
-
-/// Constructor for [`Rotation`] that's relative to previous value using currying.
-pub fn rotation_to(to: Quat) -> impl Fn(&mut Quat) -> Rotation {
-    move |state| {
-        let start = *state;
-        let end = to;
-        *state = to;
-        rotation(start, end)
-    }
-}
-
-/// Constructor for [`Rotation`] that's relative to previous value using currying.
-pub fn rotation_by(by: Quat) -> impl Fn(&mut Quat) -> Rotation {
-    move |state| {
-        let start = *state;
-        let end = *state + by;
-        *state = state.mul_quat(by);
-        rotation(start, end)
-    }
-}
-
-/// [`Interpolator`] for [`Transform`]'s scale
-#[derive(Debug, Default, Clone, PartialEq, Reflect)]
-// #[reflect(InterpolatorTransform)]
-pub struct Scale {
-    #[allow(missing_docs)]
-    pub start: Vec3,
-    #[allow(missing_docs)]
-    pub end: Vec3,
-}
-impl Interpolator for Scale {
-    type Item = Transform;
-
-    fn interpolate(&self, item: &mut Self::Item, value: f32) {
-        item.scale = self.start.lerp(self.end, value);
-    }
-}
-
-/// Constructor for [`Scale`]
-pub fn scale(start: Vec3, end: Vec3) -> Scale {
-    Scale { start, end }
-}
-
-/// Constructor for [`Scale`] that's relative to previous value using currying.
-pub fn scale_to(to: Vec3) -> impl Fn(&mut Vec3) -> Scale {
-    move |state| {
-        let start = *state;
-        let end = to;
-        *state = to;
-        scale(start, end)
-    }
-}
-
-/// Constructor for [`Scale`] that's relative to previous value using currying.
-pub fn scale_by(by: Vec3) -> impl Fn(&mut Vec3) -> Scale {
-    move |state| {
-        let start = *state;
-        let end = *state + by;
-        *state += by;
-        scale(start, end)
-    }
-}
-
-/// [`Interpolator`] for [`Transform`]'s rotation at Z axis.
-/// Usually used for 2D rotation.
-#[derive(Debug, Default, Clone, PartialEq, Reflect)]
-// #[reflect(InterpolatorTransform)]
-pub struct AngleZ {
-    #[allow(missing_docs)]
-    pub start: f32,
-    #[allow(missing_docs)]
-    pub end: f32,
-}
-impl Interpolator for AngleZ {
-    type Item = Transform;
-
-    fn interpolate(&self, item: &mut Self::Item, value: f32) {
-        let angle = (self.end - self.start).mul_add(value, self.start);
-        item.rotation = Quat::from_rotation_z(angle);
-    }
-}
-
-/// Constructor for [`AngleZ`]
-pub fn angle_z(start: f32, end: f32) -> AngleZ {
-    AngleZ { start, end }
-}
-
-/// Constructor for [`AngleZ`] that's relative to previous value using currying.
-pub fn angle_z_to(to: f32) -> impl Fn(&mut f32) -> AngleZ {
-    move |state| {
-        let start = *state;
-        let end = to;
-        *state = to;
-        angle_z(start, end)
-    }
-}
-
-/// Constructor for [`AngleZ`] that's relative to previous value using currying.
-pub fn angle_z_by(by: f32) -> impl Fn(&mut f32) -> AngleZ {
-    move |state| {
-        let start = *state;
-        let end = *state + by;
-        *state += by;
-        angle_z(start, end)
-    }
-}
-
-// #[cfg(feature = "bevy_sprite")]
-// type ReflectInterpolatorSprite = ReflectInterpolator<Sprite>;
-
-/// [`Interpolator`] for [`Sprite`]'s color
-#[cfg(feature = "bevy_sprite")]
-#[derive(Debug, Default, Clone, PartialEq, Reflect)]
-// #[reflect(InterpolatorSprite)]
-pub struct SpriteColor {
-    #[allow(missing_docs)]
-    pub start: Color,
-    #[allow(missing_docs)]
-    pub end: Color,
-}
-
-#[cfg(feature = "bevy_sprite")]
-impl Interpolator for SpriteColor {
-    type Item = Sprite;
-
-    fn interpolate(&self, item: &mut Self::Item, value: f32) {
-        item.color = color_lerp(self.start, self.end, value)
-    }
-}
-
-/// Constructor for [`SpriteColor`]
-#[cfg(feature = "bevy_sprite")]
-pub fn sprite_color(start: Color, end: Color) -> SpriteColor {
-    SpriteColor { start, end }
-}
-
-/// Constructor for [`SpriteColor`] that's relative to previous value using currying.
-#[cfg(feature = "bevy_sprite")]
-pub fn sprite_color_to(to: Color) -> impl Fn(&mut Color) -> SpriteColor {
-    move |state| {
-        let start = *state;
-        let end = to;
-        *state = to;
-        sprite_color(start, end)
-    }
-}
-
-// #[cfg(feature = "bevy_sprite")]
-// type ReflectInterpolatorColorMaterial =
-//     ReflectInterpolator<bevy::sprite::ColorMaterial>;
-
-/// [`Interpolator`] for [`Sprite`]'s [`ColorMaterial`]
-#[cfg(feature = "bevy_sprite")]
-#[derive(Debug, Default, Clone, PartialEq, Reflect)]
-// #[reflect(InterpolatorColorMaterial)]
-pub struct ColorMaterial {
-    #[allow(missing_docs)]
-    pub start: Color,
-    #[allow(missing_docs)]
-    pub end: Color,
-}
-
-#[cfg(feature = "bevy_sprite")]
-impl Interpolator for ColorMaterial {
-    type Item = bevy::sprite::ColorMaterial;
-
-    fn interpolate(&self, item: &mut Self::Item, value: f32) {
-        item.color = color_lerp(self.start, self.end, value);
-    }
-}
-
-/// Constructor for [`ColorMaterial`]
-#[cfg(feature = "bevy_sprite")]
-pub fn color_material(start: Color, end: Color) -> ColorMaterial {
-    ColorMaterial { start, end }
-}
-
-/// Constructor for [`ColorMaterial`] that's relative to previous value using currying.
-#[cfg(feature = "bevy_sprite")]
-pub fn color_material_to(to: Color) -> impl Fn(&mut Color) -> ColorMaterial {
-    move |state| {
-        let start = *state;
-        let end = to;
-        *state = to;
-        color_material(start, end)
     }
 }
